add login flow and registration user and tenant

+++ backend/core/model/document_set.go
@@ -5,7 +5,7 @@ import "github.com/google/uuid"
 type (
 	DocumentSet struct {
 		tableName   struct{}  `pg:"document_sets"`
-		ID          int       `json:"id,omitempty"`
+		ID          int64     `json:"id,omitempty"`
 		UserID      uuid.UUID `json:"user_id,omitempty"`
 		Name        string    `json:"name,omitempty"`
 		Description string    `json:"description,omitempty"`
@@ -14,9 +14,9 @@ type (
 
 	DocumentSetConnectorPair struct {
 		tableName     struct{} `pg:"document_set_connector_pairs"`
-		ID            int      `json:"id,omitempty"`
-		DocumentSetID int      `json:"document_set_id,omitempty"`
-		ConnectorID   int      `json:"connector_id,omitempty"`
+		ID            int64    `json:"id,omitempty"`
+		DocumentSetID int64    `json:"document_set_id,omitempty"`
+		ConnectorID   int64    `json:"connector_id,omitempty"`
 		IsCurrent     bool     `json:"is_current,omitempty"`
 	}
 )
